requirements.txt
Pillow
Flask
requests
pywebview